Add files via upload
The goal of this project is to implement an efficient path planning algorithm for robot navigation. We'll use the A* (A-star) algorithm, which is an informed search algorithm, to find the optimal path from a starting point to a target point on a grid map. The grid map will contain obstacles that the robot must avoid.